+++ ProdManager/routes/incident/forms.py
@@ -41,7 +41,7 @@ class IncidentCreateForm(CustomForm):
   external_link = URLField(
     name='external_link',
     label=lang.get("table_column_external_link").capitalize(),
-    validators=[Optional(), URL()]
+    validators=[Optional(), URL(require_tld=False)]
   )
   severity = SelectField(
     name='severity',